Rearranged functions in the headers and source
