+++ .gitignore
@@ -4,3 +4,4 @@ debug*
 externals/*
 *.so
 *__pycache__*
+rel_with_dbg_info/
