+++ include/utility.hpp
@@ -90,26 +90,3 @@ inline int right(int L, Edge e) // works only when horizontal
 }
 
 
-
-
-constexpr int to_vertex_index(int L, const int row, const int col)
-{
-	return (((row + L) % L)) * L + (col + L) % L;
-}
-
-inline std::tuple<int, int> vertex_to_coord(const int L, const int vidx)
-{
-	return std::make_tuple(vidx / L, vidx % L);
-}
-
-int decoder_edge_to_qubit_idx(const int L, Edge e, ErrorType error_type);
-
-void add_corrections(const int L, const std::vector<Edge>& corrections, 
-		std::vector<int>& error, ErrorType error_type);
-
-bool logical_error(const int L, const std::vector<int>& error, ErrorType error_type);
-
-
-Edge to_edge(const int L, int edge_index);
-
-
